Remove <span> from title/banner
They were pushing out of the containing element <body> meaning that on
an iPhone it was horizontally scrolling. Instead of faffing around for
ages trying to fix this just use spaces rather than spans for the same
thing anyway. Troubleshooting the CSS is a PITA not just cos it is CSS
but because responsive mode in browsers works fine, it is only when on
the physical device that the problem shows itself.